requirements.txt
six
paramiko
pywin32
wmi
iniparse
netifaces
pybootd
psutil
oslo.utils
PySide
